リファクタリング: 装備ドメインを equipment_manager へ分離＋分解/限界突破を返値メッセージ化
party_manager の god module 化を解消する第一歩。装備系（作成・装着・分解・
限界突破）を equipment_manager.py へ切り出した。

- 分解/限界突破を print() ではなく (success, message) を返す形に変更。
  これにより equipment_service は捏造文言をやめ、ドメインの実メッセージ
  （「○○を分解して△△を手に入れた」等）をそのまま flash 表示できる。
- Player ファサードの装備系委譲先を equipment_manager に変更（公開メソッド名は不変）。
- party_manager から装備系の関数と未使用 import を除去。
- 契約変更に伴い test_equipment_synthesis をタプル分解へ更新。

テスト85件グリーン。挙動は不変（メッセージがダミーから実物に変わるのみ）。
