+++ backend/tests/test_equipment_synthesis.py
@@ -30,8 +30,9 @@ class EquipmentSynthesisTests(unittest.TestCase):
         player = Player('Tester', user_id=self.user_id)
         equip = EquipmentInstance(base_item=BRONZE_SWORD, title=None)
         player.equipment_inventory.append(equip)
-        success = player.disassemble_equipment(equip.instance_id)
+        success, msg = player.disassemble_equipment(equip.instance_id)
         self.assertTrue(success)
+        self.assertIsInstance(msg, str)
         self.assertEqual(len(player.equipment_inventory), 0)
         self.assertEqual(len(player.items), 1)
         self.assertEqual(player.items[0].item_id, 'weapon_core_common')
@@ -39,7 +40,9 @@ class EquipmentSynthesisTests(unittest.TestCase):
     def test_limit_break_fails_without_material_and_keeps_item(self):
         player = Player('Tester', user_id=self.user_id)
         player.equipment_inventory.append(BRONZE_SWORD)
-        self.assertFalse(player.limit_break_equipment(BRONZE_SWORD.equip_id))
+        ok, msg = player.limit_break_equipment(BRONZE_SWORD.equip_id)
+        self.assertFalse(ok)
+        self.assertIsInstance(msg, str)
         self.assertEqual(len(player.equipment_inventory), 1)
         self.assertIs(player.equipment_inventory[0], BRONZE_SWORD)
         self.assertNotIsInstance(player.equipment_inventory[0], EquipmentInstance)
@@ -49,7 +52,8 @@ class EquipmentSynthesisTests(unittest.TestCase):
         equip = EquipmentInstance(base_item=BRONZE_SWORD, title=None)
         player.equipment_inventory.append(equip)
         player.items.extend([ALL_ITEMS['weapon_core_common'] for _ in range(5)])
-        self.assertTrue(player.limit_break_equipment(equip.instance_id))
+        ok, _ = player.limit_break_equipment(equip.instance_id)
+        self.assertTrue(ok)
         save_manager.save_game(player, self.db_path, user_id=self.user_id)
         loaded = save_manager.load_game(self.db_path, user_id=self.user_id)
         self.assertEqual(loaded.equipment_inventory[0].synthesis_rank, 1)
@@ -64,9 +68,9 @@ class EquipmentSynthesisTests(unittest.TestCase):
         self.assertEqual(equip.sub_stat_slots, initial_slots)
         player.items.extend([ALL_ITEMS['weapon_core_common'] for _ in range(20)])
         player.items.extend([ALL_ITEMS['weapon_core_rare'] for _ in range(10)])
-        self.assertTrue(player.limit_break_equipment(equip.instance_id))
-        self.assertTrue(player.limit_break_equipment(equip.instance_id))
-        self.assertTrue(player.limit_break_equipment(equip.instance_id))
+        for _ in range(3):
+            ok, _ = player.limit_break_equipment(equip.instance_id)
+            self.assertTrue(ok)
         self.assertLessEqual(
             len(equip.random_bonuses.get('sub_stats', [])), equip.sub_stat_slots
         )
